essential-code-backup/apps/backend-rag/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1  # For BaseSettings in config.py
email-validator==2.1.0  # Required for Pydantic EmailStr validation

# ChromaDB and embeddings
chromadb==1.1.0
# sentence-transformers==3.2.1  # REMOVED: Using OpenAI embeddings, not local models (saves ~5GB)
openai==1.57.4  # For text-embedding-3-small (1536-dim) and AI clients

# Machine Learning (for query clustering)
scikit-learn==1.5.2
numpy==1.26.4

# PyTorch CPU-only - NOT NEEDED (we use OpenAI embeddings)
# sentence-transformers removed, no need for PyTorch

# Anthropic API
anthropic==0.39.0

# HTTP client
httpx==0.27.2
requests==2.32.3  # For Qdrant REST API client

# Async support
tenacity==9.0.0

# Logging
loguru==0.7.2

# Environment variables
python-dotenv==1.0.1

# AWS SDK for Cloudflare R2 (S3-compatible storage)
boto3==1.35.0

# PostgreSQL drivers (Fly.io)
asyncpg==0.30.0
psycopg2-binary==2.9.10  # For migrations (sync driver)

# Server-Sent Events for streaming responses
sse-starlette==1.8.2

# Redis for caching and rate limiting (optional, falls back to in-memory)
redis==5.0.1

# Notification providers (optional)
# sendgrid==6.11.0  # Uncomment when SENDGRID_API_KEY is available
# twilio==8.10.0    # Uncomment when TWILIO credentials are available

# Force rebuild: 2025-10-25-16:00 - Phase 1 deployment

# Self-Healing Agent dependencies
psutil==5.9.6
redis==5.0.1
